moved code which originally was planned to be in the CityBikeController
