+++ app/src/main/java/com/group1/movebetter/view_model/controller/MapController.kt
@@ -25,133 +25,7 @@ import kotlinx.coroutines.CoroutineScope
 
 class MapController(private val viewModelScope: CoroutineScope, private val repository: Repository) {
 
-    private val networkFeatures: ArrayList<Feature> = ArrayList()
-    private var currentNetwork: CityBikesNetwork? = null
-
-    fun createBikeNetworkList(networkSource: GeoJsonSource?, cityBikes: CityBikes, cityBikeController: CityBikeController) {
-        for (network in cityBikes.networks) {
-            val feature = createBikeFeature(network.id, network.location, true)
-
-            if (network.id == closestNetwork.id) {
-                feature.addBooleanProperty("show", false)
-                cityBikeController.getNetwork(network.id)
-            }
-
-            networkFeatures.add(feature)
-        }
-
-        refreshSource(networkSource!!, networkFeatures)
-    }
-
-    private fun createBikeFeature(id: String, location: CityBikesLocation, show: Boolean): Feature {
-        val feature = Feature.fromGeometry(Point.fromLngLat(location.longitude, location.latitude))
-        feature.addStringProperty("id", id)
-        feature.addBooleanProperty("show", show)
-
-        return feature
-    }
-
-    @SuppressLint("NewApi")
-    fun exchangeNetworkWithStations(networkSource: GeoJsonSource?, stationSource: GeoJsonSource?, network: CityBikesNetwork) {
-        // Update currentNetwork
-        if (currentNetwork == null) {
-            currentNetwork = network
-        }
-
-        val condition = Predicate<Feature> { x -> x.getStringProperty("id").equals(network.id) || x.getStringProperty("id").equals(currentNetwork!!.id) }
-        networkFeatures.removeIf(condition)
-
-        // Update feature list of bike networks
-        networkFeatures.add(createBikeFeature(network.id, network.location, false))
-
-        if (currentNetwork!!.id != network.id) {
-            networkFeatures.add(createBikeFeature(currentNetwork!!.id, currentNetwork!!.location, true))
-            currentNetwork = network
-        }
-
-        refreshSource(networkSource!!, networkFeatures)
-
-        // Add stations to map
-        val featureList = ArrayList<Feature>()
-
-        for (station in network.stations) {
-            val featureStation = Feature.fromGeometry(Point.fromLngLat(station.longitude, station.latitude))
-            featureStation.addStringProperty("name", station.name)
-            featureStation.addStringProperty("id", station.id)
-            featureStation.addNumberProperty("freeBikes", station.free_bikes.amount)
-            featureStation.addNumberProperty("emptySlots", station.empty_slots.amount)
-            featureStation.addStringProperty("timestamp", station.timestamp)
-            featureStation.addStringProperty("provider", "bikes")
-
-            featureList.add(featureStation)
-        }
-
-        refreshSource(stationSource!!, featureList)
-    }
-
     fun refreshSource(source: GeoJsonSource, featureList: ArrayList<Feature>) {
         source.setGeoJson(FeatureCollection.fromFeatures(featureList))
-    }
-
-    // nearest-network logic
-    private lateinit var fusedLocationClient: FusedLocationProviderClient
-    private var permissionsManager: PermissionsManager? = null
-
-    lateinit var closestNetwork: CityBikesNetworks
-    var currentLocation: Location = initLocation()
-
-    private fun initLocation() : Location {
-
-        // Setting default location to Berlin
-
-        val location = Location(LocationManager.GPS_PROVIDER)
-        location.longitude = 13.404954
-        location.latitude = 52.520008
-        return location
-    }
-
-    fun getNearestNetwork(res: CityBikes) {
-
-        // Haversine formular (see more here: https://en.wikipedia.org/wiki/Haversine_formula)
-
-        val earthRadius = 6371
-        val distances = ArrayList<Double>()
-        val distanceNetworkMap = HashMap<Double, CityBikesNetworks>()
-        for (network in res.networks) {
-            val networkLocation: CityBikesLocation = network.location
-            val distanceLatitude = degreeToRadial(abs(currentLocation.latitude) - abs(networkLocation.latitude))
-            val distanceLongitude = degreeToRadial(abs(currentLocation.longitude) - abs(networkLocation.longitude))
-
-            val a = sin(distanceLatitude / 2) * sin(distanceLatitude / 2) +
-                    cos(degreeToRadial(abs(networkLocation.latitude))) * cos(degreeToRadial(abs(currentLocation.latitude))) *
-                    sin(distanceLongitude / 2) * sin(distanceLongitude / 2)
-
-            val c = 2 * atan2(sqrt(a), sqrt(1 - a));
-            val d = earthRadius * c; // Distance in km
-            distances.add(d)
-            distanceNetworkMap[d] = network
-        }
-        val minDistance: Double? = distances.minByOrNull { it }
-        closestNetwork = distanceNetworkMap[minDistance]!!
-    }
-
-    private fun degreeToRadial(degree: Double): Double {
-        return degree * (Math.PI/180)
-    }
-
-    @SuppressLint("MissingPermission")
-    fun getCurrentLocation(activity: FragmentActivity, context: Context?, mapFragment: MapFragment) {
-        fusedLocationClient = LocationServices.getFusedLocationProviderClient(activity)
-
-        if (!PermissionsManager.areLocationPermissionsGranted(context)) {
-            permissionsManager = PermissionsManager(mapFragment)
-            permissionsManager!!.requestLocationPermissions(activity)
-        }
-
-        fusedLocationClient.lastLocation.addOnSuccessListener { location: Location? ->
-            if (location != null) {
-                currentLocation = location
-            }
-        }
     }
 }
